telgrambot/requirements.txt
aiogram
loguru
python-dotenv
psycopg2
sqlalchemy
requests
